refactor(ai-blog-writer): remove orphan prompt config
