--- apps/ai-blog-writer/apps/frontend/src/features/imageRecreationPrompts/types.ts
@@ -1,358 +1,0 @@
-export type SceneCategoryId =
-  | 'landscape-only'
-  | 'scenic-viewpoint'
-  | 'tourist-landmark'
-  | 'city-street-scene'
-  | 'street-art-mural'
-  | 'market-food-stall'
-  | 'restaurant-interior'
-  | 'restaurant-exterior'
-  | 'cafe-interior'
-  | 'cafe-exterior'
-  | 'cafe-restaurant-scene'
-  | 'plated-food-close-up'
-  | 'tabletop-food-spread'
-  | 'coffee-drinks-table'
-  | 'bakery-pastry-display'
-  | 'bar-cocktail-scene'
-  | 'hotel-lobby-interior'
-  | 'hotel-room-interior'
-  | 'spa-wellness-interior'
-  | 'poolside-resort-scene'
-  | 'boutique-shopfront'
-  | 'rooftop-terrace-view'
-  | 'nightlife-neon-scene'
-  | 'architecture-exterior'
-  | 'architecture-interior'
-  | 'museum-gallery-interior'
-  | 'portrait'
-  | 'couple-friends-photo'
-  | 'group-photo'
-  | 'lifestyle-candid-people'
-  | 'product-object'
-  | 'nature-wilderness'
-  | 'beach-coastal'
-  | 'mountain-hiking'
-  | 'desert-rock-formations'
-
-export type PeoplePresenceId =
-  | 'no-people'
-  | 'one-person'
-  | 'two-people'
-  | 'small-group'
-  | 'spread-out-crowd'
-  | 'dense-crowd'
-
-export type PeopleHandlingId =
-  | 'match-reference-people-handling'
-  | 'keep-every-person-as-is'
-  | 'keep-same-people-small-natural-changes'
-  | 'keep-same-people-small-repositioning'
-  | 'reduce-a-few-people'
-  | 'remove-all-people'
-  | 'change-every-person-and-reshuffle'
-  | 'reshuffle-and-add-people-naturally'
-  | 'people-secondary-environment-primary'
-  | 'environment-dominant-with-people'
-  | 'custom-people-handling'
-
-export type CrowdCharacterId =
-  | 'match-reference-crowd'
-  | 'international-tourist-mix'
-  | 'locals-dominant'
-  | 'mixed-age-travelers'
-  | 'family-heavy-travelers'
-  | 'adult-travelers'
-  | 'backpacker-travel-crowd'
-  | 'stylish-city-weekend-crowd'
-  | 'understated-neutral-crowd'
-
-export type PrimarySubjectEmphasisId =
-  | 'match-reference-subject-balance'
-  | 'environment-first'
-  | 'landmark-first'
-  | 'person-first'
-  | 'balanced-scene'
-
-export type PeopleStrategyId =
-  | 'match-source'
-  | 'keep'
-  | 'reduce'
-  | 'remove'
-  | 'recast-or-add'
-
-export type CameraPresetId =
-  | 'sony-a7r-v'
-  | 'sony-a7-iv'
-  | 'sony-a1'
-  | 'sony-a9-iii'
-  | 'canon-r5'
-  | 'canon-r6-mark-ii'
-  | 'canon-r1'
-  | 'canon-r3'
-  | 'nikon-z8'
-  | 'nikon-z9'
-  | 'nikon-zf'
-  | 'fujifilm-gfx-100s'
-  | 'hasselblad-x2d-100c'
-  | 'fujifilm-x-t5'
-  | 'fujifilm-x100vi'
-  | 'leica-q3'
-  | 'leica-sl3'
-  | 'leica-m11'
-  | 'contax-t2'
-  | 'yashica-t4'
-  | 'leica-m6'
-  | 'hasselblad-500cm'
-  | 'mamiya-7'
-  | 'pentax-67'
-  | 'canon-ae-1'
-  | 'nikon-fm2'
-  | 'olympus-om-1'
-  | 'minolta-cle'
-  | 'polaroid-sx-70'
-
-export type LensPresetId =
-  | '20mm-f1-8'
-  | '24mm-f1-4'
-  | '24mm-f1-8'
-  | '28mm-f2'
-  | '28-70mm-f2'
-  | '35mm-f1-8'
-  | '35mm-f1-4'
-  | '50mm-f1-2'
-  | '50mm-f1-4'
-  | '50mm-f1-8'
-  | '90mm-f2'
-  | '85mm-f1-8'
-  | '85mm-f1-2'
-  | '135mm-f1-8'
-  | '14-24mm-f2-8'
-  | '70-200mm-f2-8'
-  | '100-400mm-f4-5-5-6'
-  | '24-70mm-f2-8'
-  | '16-35mm-f2-8'
-  | '21mm-vintage-ultra-wide'
-  | '35mm-vintage-rangefinder'
-  | '40mm-vintage-pancake'
-  | '50mm-vintage-fast-prime'
-  | '85mm-vintage-portrait'
-  | 'soft-vintage-film-lens'
-  | 'swirly-vintage-portrait-lens'
-  | 'anamorphic-vintage-inspired'
-  | 'classic-medium-format-rendering'
-  | '45mm-equivalent-medium-format'
-
-export type CaptureStyleId =
-  | 'match-reference-style'
-  | 'editorial'
-  | 'natural-documentary'
-  | 'moody-documentary'
-  | 'luxury-campaign'
-  | 'hospitality-editorial'
-  | 'travel-photography'
-  | 'street-photography'
-  | 'cinematic-still'
-  | 'fine-art-landscape'
-  | 'filmic-vintage'
-  | 'real-estate-architecture-clean'
-  | 'interior-design-editorial'
-  | 'food-editorial'
-  | 'commercial-product'
-  | 'brand-social-content'
-  | 'casual-candid'
-
-export type ShotPerspectiveId =
-  | 'match-reference-viewpoint'
-  | 'eye-level-natural'
-  | 'low-angle-upward'
-  | 'ground-level-dramatic'
-  | 'high-angle-downward'
-  | 'elevated-overlook'
-  | 'birds-eye-overhead'
-  | 'drone-oblique'
-  | 'straight-on-frontal'
-  | 'three-quarter-oblique'
-  | 'side-profile-angle'
-  | 'tilted-dynamic'
-  | 'foreground-led-wide'
-  | 'telephoto-observer'
-
-export type FilterLookId =
-  | 'neutral-no-filter'
-  | 'iphone-natural'
-  | 'iphone-vivid'
-  | 'fujifilm-classic-chrome'
-  | 'fujifilm-nostalgic-neg'
-  | 'fujifilm-reala-ace'
-  | 'fujifilm-pro-400h'
-  | 'kodak-portra-400'
-  | 'kodak-portra-800'
-  | 'kodak-gold-200'
-  | 'kodak-ektar-100'
-  | 'kodak-ultramax-400'
-  | 'cinestill-800t'
-  | 'kodachrome-64'
-  | 'agfa-vista-200'
-  | 'modern-vintage-soft-warm'
-  | 'modern-vintage-faded-clean'
-  | 'clean-analog-color'
-  | 'soft-editorial-nostalgia'
-  | 'subtle-analog-grain'
-  | 'faded-print-vintage'
-  | 'faded-disposable-film'
-  | 'dusty-postcard-vintage'
-  | 'sun-faded-travel-print'
-  | 'retro-chrome-slide'
-  | 'warm-archive-postcard'
-  | 'expired-color-negative'
-  | 'ilford-hp5-bw'
-  | 'kodak-tri-x-400'
-  | 'leica-natural'
-
-export type LightingId =
-  | 'match-reference-lighting'
-  | 'clear-bright-midday-sun'
-  | 'clear-morning-sun'
-  | 'soft-morning-light'
-  | 'pastel-sunrise-light'
-  | 'sun-through-thin-cloud'
-  | 'broken-cloud-sun'
-  | 'golden-hour'
-  | 'sunset-glow'
-  | 'after-rain-sunbreaks'
-  | 'blue-hour'
-  | 'after-sunset-twilight'
-  | 'overcast-soft-light'
-  | 'bright-high-overcast'
-  | 'diffused-cloudy-daylight'
-  | 'heavy-overcast-dark-clouds'
-  | 'dramatic-storm-light'
-  | 'hazy-afternoon-light'
-  | 'backlit-sunlight'
-  | 'indoor-general'
-  | 'indoor-production'
-  | 'window-light'
-  | 'night-city-lights'
-  | 'mixed-urban-lighting'
-  | 'flat-neutral-daylight'
-
-export type PreservationStrengthId =
-  | 'match-reference-preservation'
-  | 'strict'
-  | 'balanced'
-  | 'flexible'
-
-export type AllowedVariationId =
-  | 'no-variation'
-  | 'small-environmental-cleanup'
-  | 'small-wardrobe-changes'
-  | 'small-positional-shifts'
-  | 'minor-secondary-detail-changes'
-
-export type EnvironmentEnhancementId =
-  | 'match-reference-environment'
-  | 'minimal'
-  | 'moderate-realism-boost'
-  | 'strong-realism-boost'
-
-export type FluxModelId =
-  | 'flux-2-max'
-  | 'flux-2-pro-preview'
-  | 'flux-2-pro'
-  | 'flux-2-flex'
-
-export type FluxSafetyToleranceId = '0' | '1' | '2' | '3' | '4' | '5'
-
-export type PromptPresetId =
-  | 'famous-landmark-no-people'
-  | 'desert-landscape-editorial'
-  | 'famous-landmark-sparse-people'
-  | 'city-square-blue-hour'
-  | 'mountain-viewpoint'
-  | 'couple-travel-photo'
-  | 'vintage-street-scene'
-  | 'custom'
-
-export interface PromptBlock {
-  id:
-    | 'reference-anchoring'
-    | 'scene-preservation'
-    | 'people-preservation'
-    | 'camera-lens-realism'
-    | 'lighting-description'
-    | 'style-description'
-    | 'environment-realism'
-    | 'user-guidance'
-    | 'negative-instructions'
-  title: string
-  text: string
-}
-
-export interface PromptBuildResult {
-  blocks: PromptBlock[]
-  finalPrompt: string
-}
-
-export interface ImageRecreationFormState {
-  presetId: PromptPresetId
-  sceneCategory: SceneCategoryId
-  referenceHasPeople: boolean
-  peopleStrategy: PeopleStrategyId
-  peopleOverrideText: string
-  cameraPreset: CameraPresetId
-  lensPreset: LensPresetId
-  captureStyle: CaptureStyleId
-  shotPerspective: ShotPerspectiveId
-  centerMainSubject: boolean
-  filterLook: FilterLookId
-  lighting: LightingId
-  preservationStrength: PreservationStrengthId
-  environmentEnhancement: EnvironmentEnhancementId
-  modelId: FluxModelId
-  safetyTolerance: FluxSafetyToleranceId
-  enablePromptUpsampling: boolean
-  seedValue: string
-  creativeDirection: string
-}
-
-export interface SelectOption<TId extends string> {
-  id: TId
-  label: string
-  description: string
-  prompt: string
-  disabled?: boolean
-}
-
-export interface OptionGroup<TId extends string> {
-  label: string
-  options: SelectOption<TId>[]
-}
-
-export interface SceneCategoryOption extends SelectOption<SceneCategoryId> {
-  helperText: string
-  recommendedPeoplePresence: PeoplePresenceId
-}
-
-export interface PromptPreset {
-  id: Exclude<PromptPresetId, 'custom'>
-  label: string
-  description: string
-  values: Omit<
-    ImageRecreationFormState,
-    | 'presetId'
-    | 'referenceHasPeople'
-    | 'centerMainSubject'
-    | 'modelId'
-    | 'safetyTolerance'
-    | 'enablePromptUpsampling'
-    | 'seedValue'
-    | 'creativeDirection'
-    // Operator-entered at runtime; createFormStateFromPreset always seeds it empty.
-    | 'peopleOverrideText'
-  > & {
-    referenceHasPeople?: boolean
-    centerMainSubject?: boolean
-    creativeDirection?: string
-  }
-}
